pyproject.toml
[project]
name = "snazel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.8.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
# snazel = "snazel.cli.snazel:run"
snazel = "snazel.__main__:main"

[tool.uv]
package = true
dev-dependencies = [
    "ruff>=0.9.1",
]

[tool.ruff]
line-length = 80
indent-width = 4

target-version = "py312"
